patch datasources
