+++ docs/platypus2/monitoring/promstack/config-map_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) 2023 Volvo Car Corporation
+// SPDX-License-Identifier: Apache-2.0
+
 package promstack
 
 import (
