+++ docs/platypus2/monitoring/vmk8s/dashboards/vmsingle.json
@@ -5603,7 +5603,7 @@
     ]
   },
   "timezone": "",
-  "title": "VictoriaMetrics",
+  "title": "VictoriaMetrics - single",
   "uid": "wNf0q_kZk",
   "version": 1,
   "weekStart": ""
